weather_app/requirements.txt
requests==2.31
python-dotenv==1.1.0
